+++ CHANGES.rst
@@ -30,6 +30,7 @@ Release History
 - Correctly handle input feeds when n_steps > step_blocks
 - Detect cycles in transitive planner
 - Fix bug in uneven step_blocks rounding
+- Fix bug in Simulator.print_params
 
 
 0.3.0 (April 25, 2017)
@@ -87,6 +88,7 @@ Release History
 ----------------------
 
 Initial release of TensorFlow-based NengoDL
+
 
 0.1.0 (June 12, 2016)
 ---------------------